+++ ide-common/src/main/kotlin/org/digma/intellij/plugin/auth/LoginResult.kt
@@ -1,0 +1,3 @@
+package org.digma.intellij.plugin.auth
+
+data class LoginResult(val isSuccess: Boolean, val userId: String?, val error: String?)